refactor: 유틸, 스토어 분리
