+++ src/js/store.js
@@ -1,0 +1,4 @@
+export const store = {
+  currentTab: MENU.EspressoMenu,
+  [MENU.EspressoMenu]: [],
+};
