+++ src/js/utils.js
@@ -1,0 +1,8 @@
+// 인풋창 validation
+export const isEmpty = (value) => {
+  if (value.replace(/ /g, "") === "") {
+    return true;
+  }
+
+  return false;
+};
